Units have summoning sickness [20 min]
The turn they are played, units are afflicted with summoning sickness,
and can't attack. Summoning Sickness no longer affects the unit at the
end of the turn.

Note: Most of the time spent on this one was just me wrapping my head
around the commands queue and how it was implemented here. If were not
including that knock it down to 10 minutes.